+++ packages/riverpod/lib/src/framework/auto_dispose.dart
@@ -59,19 +59,32 @@ abstract class AutoDisposeProviderBase<State> extends ProviderBase<State> {
     required Family? from,
     required Object? argument,
     required this.cacheTime,
+    required this.disposeDelay,
   }) : super(name: name, from: from, argument: argument);
 
   /// {@template riverpod.cache_time}
   /// The minimum amount of time before an `autoDispose` provider can be
   /// disposed if not listened after the last value change.
   ///
-  /// If the provider rebuilds (such as when using `ref.watch` or `ref.refresh`),
-  /// the timer will be refreshed.
+  /// If the provider rebuilds (such as when using `ref.watch` or `ref.refresh`)
+  /// or emits a new value, the timer will be refreshed.
   ///
   /// If null, use the nearest ancestor [ProviderContainer]'s [cacheTime].
   /// If no ancestor is found, fallbacks to [Duration.zero].
   /// {@endtemplate}
   final Duration? cacheTime;
+
+  /// {@template riverpod.dispose_delay}
+  /// The amount of time before a provider is disposed after its last listener
+  /// is removed.
+  ///
+  /// If a new listener is added within that duration, the provider will not be
+  /// disposed.
+  ///
+  /// If null, use the nearest ancestor [ProviderContainer]'s [disposeDelay].
+  /// If no ancestor is found, fallbacks to [Duration.zero].
+  /// {@endtemplate}
+  final Duration? disposeDelay;
 
   @override
   State create(AutoDisposeRef ref);
@@ -95,6 +108,10 @@ abstract class AutoDisposeProviderElementBase<State>
   late final _cacheTime =
       (provider as AutoDisposeProviderBase).cacheTime ?? _container.cacheTime;
 
+  late final _disposeDelay =
+      (provider as AutoDisposeProviderBase).disposeDelay ??
+          _container.disposeDelay;
+
   @Deprecated('Use `keepAlive()` instead')
   @override
   bool get maintainState => _maintainState;
@@ -104,7 +121,10 @@ abstract class AutoDisposeProviderElementBase<State>
     if (!value) mayNeedDispose();
   }
 
-  Timer? _timer;
+  Timer? _cacheTimer;
+
+  KeepAliveLink? _disposeDelayLink;
+  Timer? _disposeDelayTimer;
 
   @override
   KeepAliveLink keepAlive() {
@@ -130,6 +150,42 @@ abstract class AutoDisposeProviderElementBase<State>
   @override
   void _buildState() {
     super._buildState();
+
+    if (_disposeDelay != Duration.zero) {
+      // TODO timer should not refresh when provider rebuilds
+      // TODO adding a new listener cancels the timer
+
+      // If we have a previous link and the provider was refreshed,
+      // then that link was removed. So it's safe to replace the link with a
+      // new one
+      assert(
+        _disposeDelayLink == null ||
+            !_keepAliveLinks.contains(_disposeDelayLink),
+        'Bad state',
+      );
+      _disposeDelayLink = keepAlive();
+
+      onCancel(() {
+        assert(_disposeDelayLink != null, 'Bad state');
+        assert(_disposeDelayTimer == null, 'Bad state');
+        _disposeDelayTimer = Timer(_disposeDelay, () {
+          _disposeDelayTimer = null;
+          _disposeDelayLink!.close();
+          _disposeDelayLink = null;
+        });
+      });
+
+      onResume(() {
+        _disposeDelayTimer?.cancel();
+        _disposeDelayTimer = null;
+        _disposeDelayLink ??= keepAlive();
+      });
+
+      // We don't cancel the timer on dispose
+      // as we voluntarily want the disposeDelay logic to be independent from
+      // when the last value was emitted.
+    }
+
     if (_cacheTime != Duration.zero) {
       // Safe to have as a local variable since links are cleared
       // on rebuild
@@ -137,12 +193,12 @@ abstract class AutoDisposeProviderElementBase<State>
 
       listenSelf((previous, next) {
         link ??= keepAlive();
-        _timer?.cancel();
+        _cacheTimer?.cancel();
 
-        _timer = Timer(_cacheTime, () {
+        _cacheTimer = Timer(_cacheTime, () {
           link!.close();
           link = null;
-          _timer = null;
+          _cacheTimer = null;
 
           // will always be initialized so `!` is safe
           // requireState is safe because if an error is emitted, the timer
@@ -154,12 +210,12 @@ abstract class AutoDisposeProviderElementBase<State>
         });
       }, onError: (err, stack) {
         link ??= keepAlive();
-        _timer?.cancel();
+        _cacheTimer?.cancel();
 
-        _timer = Timer(_cacheTime, () {
+        _cacheTimer = Timer(_cacheTime, () {
           link!.close();
           link = null;
-          _timer = null;
+          _cacheTimer = null;
         });
       });